Precompiled header + configured to compile as library
