+++ BNN.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.6.33829.357
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BNN", "BNN\BNN.vcxproj", "{40067B61-DD89-455A-8A65-B56756D83935}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Example1", "Example1\Example1.vcxproj", "{2166F13D-82CA-4DCA-868A-A03CE321D6B5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -21,6 +23,14 @@ Global
 		{40067B61-DD89-455A-8A65-B56756D83935}.Release|x64.Build.0 = Release|x64
 		{40067B61-DD89-455A-8A65-B56756D83935}.Release|x86.ActiveCfg = Release|Win32
 		{40067B61-DD89-455A-8A65-B56756D83935}.Release|x86.Build.0 = Release|Win32
+		{2166F13D-82CA-4DCA-868A-A03CE321D6B5}.Debug|x64.ActiveCfg = Debug|x64
+		{2166F13D-82CA-4DCA-868A-A03CE321D6B5}.Debug|x64.Build.0 = Debug|x64
+		{2166F13D-82CA-4DCA-868A-A03CE321D6B5}.Debug|x86.ActiveCfg = Debug|Win32
+		{2166F13D-82CA-4DCA-868A-A03CE321D6B5}.Debug|x86.Build.0 = Debug|Win32
+		{2166F13D-82CA-4DCA-868A-A03CE321D6B5}.Release|x64.ActiveCfg = Release|x64
+		{2166F13D-82CA-4DCA-868A-A03CE321D6B5}.Release|x64.Build.0 = Release|x64
+		{2166F13D-82CA-4DCA-868A-A03CE321D6B5}.Release|x86.ActiveCfg = Release|Win32
+		{2166F13D-82CA-4DCA-868A-A03CE321D6B5}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
