+++ BNN/BNN.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "NNet/NNet.h"
+#include "NNet/NNet_samples.h"
+#include "Image/Image.h"
