+++ Example1/pch.h
@@ -1,0 +1,2 @@
+#pragma once
+#include <BNN.h>
